Fixed bugs and formatted code

+++ VC4 Launcher/src/vc4/launcher/task/RemovePackageTask.java
@@ -27,29 +27,30 @@ public class RemovePackageTask implements Task {
 			return;
 		}
 		Install instal = pack.getInstalls()[install];
-		if(!instal.canInstall(pack.getVersion())){
+		if (!instal.canInstall(pack.getVersion())) {
 			progress.setPercent(100);
 			progress.setDelete(true);
-			if(install == pack.getInstalls().length - 1) pack.setVersion(null);;
+			if (install == pack.getInstalls().length - 1) pack.setVersion(null);
+			;
 		}
 		String installPath = DirectoryLocator.getPath() + instal.getEnd();
-		if(pack.isDisabled()) installPath = installPath + ".disabled";
+		if (pack.isDisabled()) installPath = installPath + ".disabled";
 		progress.setPercent(33);
 		if (instal.getType() == InstallType.UNZIP) {
 			progress.setText("Deleting: " + instal.getEnd());
 			installPath = installPath + "/";
 			progress.setPercent(66);
 			File files = new File(installPath);
-			if(files.exists()) deleteDirectory(files);
+			if (files.exists()) deleteDirectory(files);
 			progress.setPercent(100);
-		} else if(instal.getType() == InstallType.COPY){
+		} else if (instal.getType() == InstallType.COPY) {
 			progress.setText("Deleting: " + instal.getEnd());
 			progress.setPercent(50);
 			File files = new File(installPath);
-			if(files.exists()) files.delete();
+			if (files.exists()) files.delete();
 			progress.setPercent(100);
 		}
-		if(install == pack.getInstalls().length - 1) pack.setVersion(null);
+		if (install == pack.getInstalls().length - 1) pack.setVersion(null);
 		progress.setDelete(true);
 		try {
 			pack.save();
@@ -63,20 +64,18 @@ public class RemovePackageTask implements Task {
 		return true;
 
 	}
-	
-	public boolean deleteDirectory(File dir) {
-	    if (dir.isDirectory()) {
-	        String[] children = dir.list();
-	        for (int i=0; i<children.length; i++) {
-	            boolean success = deleteDirectory(new File(dir, children[i]));
-	            if (!success) {
-	                return dir.delete();
-	            }
-	        }
-	    }
 
-	    // The directory is now empty so delete it
-	    return dir.delete();
+	public boolean deleteDirectory(File dir) {
+		if (dir.isDirectory()) {
+			String[] children = dir.list();
+			for (int i = 0; i < children.length; i++) {
+				boolean success = deleteDirectory(new File(dir, children[i]));
+				if (!success) { return dir.delete(); }
+			}
+		}
+
+		// The directory is now empty so delete it
+		return dir.delete();
 	}
 
 }
